+++ source/ui/desktop/widgets.py
@@ -8179,1399 +8179,6 @@ def button_action(button_name, button, specific_screen=''):
             utility.screen_manager.current = "CreateServerProgressScreen"
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-class TemplateButton(HoverButton):
-
-    def animate_button(self, image, color, hover_action, **kwargs):
-        image_animate = Animation(duration=0.05)
-
-        Animation(color=color, duration=0.06).start(self.title)
-        Animation(color=color, duration=0.06).start(self.subtitle)
-
-        Animation(color=color, duration=0.06).start(self.type_image.image)
-
-        if self.type_image.version_label.__class__.__name__ == "AlignLabel":
-            Animation(color=color, duration=0.06).start(self.type_image.version_label)
-        Animation(color=color, duration=0.06).start(self.type_image.type_label)
-
-        _animate_background(self, image, hover_action)
-
-        image_animate.start(self)
-
-    def resize_self(self, *args):
-
-        # Title and description
-        padding = 2.17
-        self.title.pos = (self.x + (self.title.text_size[0] / padding) - (8.3) + 30, self.y + 31)
-        self.subtitle.pos = (self.x + (self.subtitle.text_size[0] / padding) - 78, self.y + 8)
-
-
-        offset = 9.45 if self.type_image.type_label.text in ["vanilla", "paper", "purpur"]\
-            else 9.6 if self.type_image.type_label.text == "forge"\
-            else 9.35 if self.type_image.type_label.text == "craftbukkit"\
-            else 9.55
-
-
-        self.type_image.image.x = self.width + self.x - (self.type_image.image.width) - 13
-        self.type_image.image.y = self.y + ((self.height / 2) - (self.type_image.image.height / 2))
-
-
-        self.type_image.type_label.x = self.width + self.x - (self.padding_x * offset) - self.type_image.width - 83
-        self.type_image.type_label.y = self.y + (self.height * 0.05)
-
-        # Update label
-        if self.type_image.version_label.__class__.__name__ == "AlignLabel":
-            self.type_image.version_label.x = self.width + self.x - (self.padding_x * offset) - self.type_image.width - 83
-            self.type_image.version_label.y = self.y - (self.height / 3.2)
-
-        # Favorite button
-        self.customize_layout.size_hint_max = (self.size_hint_max[0], self.size_hint_max[1])
-        self.customize_layout.pos = (self.pos[0] - 6, self.pos[1] + 13)
-
-    def __init__(self, template, fade_in=0.0, **kwargs):
-        super().__init__(**kwargs)
-
-        self.template = template
-        self.border = (-5, -5, -5, -5)
-        self.color_id = [(0.05, 0.05, 0.1, 1), constants.brighten_color((0.65, 0.65, 1, 1), 0.07)]
-        self.pos_hint = {"center_x": 0.5, "center_y": 0.6}
-        self.size_hint_max = (580, 80)
-        self.id = "server_button"
-
-        self.background_normal = os.path.join(paths.ui_assets, f'{self.id}.png')
-        self.background_down = os.path.join(paths.ui_assets, f'{self.id}_click.png')
-
-
-        # Title of Server
-        self.title = Label()
-        self.title.__translate__ = False
-        self.title.id = "title"
-        self.title.halign = "left"
-        self.title.color = self.color_id[1]
-        self.title.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["medium"]}.ttf')
-        self.title.font_size = sp(25)
-        self.title.text_size = (self.size_hint_max[0] * 0.58, self.size_hint_max[1])
-        self.title.shorten = True
-        self.title.markup = True
-        self.title.shorten_from = "right"
-        self.title.max_lines = 1
-        self.title.text = template['template']['name']
-        self.add_widget(self.title)
-
-
-        # Server last modified date formatted
-        self.subtitle = Label()
-        self.subtitle.__translate__ = False
-        self.subtitle.size = (300, 30)
-        self.subtitle.id = "subtitle"
-        self.subtitle.halign = "left"
-        self.subtitle.valign = "center"
-        self.subtitle.font_size = sp(21)
-        self.subtitle.text_size = (self.size_hint_max[0] * 0.91, self.size_hint_max[1])
-        self.subtitle.shorten = True
-        self.subtitle.markup = True
-        self.subtitle.shorten_from = "right"
-        self.subtitle.max_lines = 1
-        self.subtitle.text_size[0] = 350
-        self.subtitle.color = self.color_id[1]
-        self.subtitle.default_opacity = 0.56
-        self.subtitle.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["regular"]}.ttf')
-
-        self.subtitle.text = template['template']['description']
-
-        self.subtitle.opacity = self.subtitle.default_opacity
-
-        self.add_widget(self.subtitle)
-
-
-        # Type icon and info
-        self.type_image = RelativeLayout()
-        self.type_image.width = 400
-
-        server_icon = os.path.join(paths.ui_assets, 'icons', 'big', f"{template['server']['type']}_small.png")
-        self.type_image.image = Image(source=server_icon)
-
-        self.type_image.image.allow_stretch = True
-        self.type_image.image.size_hint_max = (65, 65)
-        self.type_image.image.color = self.color_id[1]
-        self.type_image.add_widget(self.type_image.image)
-
-        def TemplateLabel():
-            template_label = AlignLabel()
-            template_label.__translate__ = False
-            template_label.halign = "right"
-            template_label.valign = "middle"
-            template_label.text_size = template_label.size
-            template_label.font_size = sp(19)
-            template_label.color = self.color_id[1]
-            template_label.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["medium"]}.ttf')
-            template_label.width = 150
-            return template_label
-
-        self.type_image.version_label = TemplateLabel()
-        self.type_image.version_label.color = self.color_id[1]
-        self.type_image.version_label.text = template['server']['version']
-        self.type_image.version_label.opacity = 0.6
-
-
-        self.type_image.type_label = TemplateLabel()
-
-        type_text = template['server']['type'].lower().replace("craft", "")
-        self.type_image.type_label.text = type_text
-        self.type_image.type_label.font_size = sp(23)
-        self.type_image.add_widget(self.type_image.version_label)
-        self.type_image.add_widget(self.type_image.type_label)
-        self.add_widget(self.type_image)
-
-
-        # Favorite button
-        self.customize_layout = RelativeLayout()
-
-        def customize_with_template(*a):
-            foundry.apply_template(self.template)
-            utility.screen_manager.current = 'CreateServerNameScreen'
-
-        def create_with_template(*a):
-            foundry.apply_template(self.template)
-            utility.screen_manager.current = 'CreateServerProgressScreen'
-
-        self.customize_button = IconButton('', {}, (0, 0), (None, None), 'settings-sharp.png', clickable=True, anchor='right', click_func=customize_with_template)
-
-        self.customize_layout.add_widget(self.customize_button)
-        self.add_widget(self.customize_layout)
-
-        self.bind(pos=self.resize_self)
-        self.bind(on_press=create_with_template)
-
-        # Animate opacity
-        if fade_in > 0:
-            self.opacity = 0
-            self.title.opacity = 0
-
-            Animation(opacity=1, duration=fade_in).start(self)
-            Animation(opacity=1, duration=fade_in).start(self.title)
-            Animation(opacity=self.subtitle.default_opacity, duration=fade_in).start(self.subtitle)
-
-    def on_enter(self, *args):
-        if not self.ignore_hover:
-            self.animate_button(image=os.path.join(paths.ui_assets, f'{self.id}_hover.png'), color=self.color_id[0], hover_action=True)
-
-    def on_leave(self, *args):
-        if not self.ignore_hover:
-            self.animate_button(image=os.path.join(paths.ui_assets, f'{self.id}.png'), color=self.color_id[1], hover_action=False)
-
-
-class RuleButton(FloatLayout):
-
-    def __setattr__(self, attr, value):
-
-        # Change attributes dynamically based on rule
-        if attr == "rule" and value:
-            self.text.text = value.rule.replace("!w", "")
-            self.change_properties(value)
-
-        super().__setattr__(attr, value)
-
-
-    # Modifies rule attributes based on rule.list_enabled status
-    def change_properties(self, rule):
-
-        # Prevent crash from spamming the back button
-        if "Acl" not in utility.screen_manager.current_screen.name:
-            return
-
-        Animation.cancel_all(self.button)
-        Animation.cancel_all(self.text)
-        Animation.cancel_all(self.icon)
-
-        self.button.id = f'rule_button{"_enabled" if rule.list_enabled else ""}'
-        self.button.background_normal = os.path.join(paths.ui_assets, f'{self.button.id}.png')
-        self.button.background_down   = os.path.join(paths.ui_assets, f'{self.button.id}_click.png')
-
-        # Change color attributes
-        if utility.screen_manager.current_screen.current_list == "ops":
-            self.color_id = self.button.color_id = [(0, 0, 0, 0.85), (0.439, 0.839, 1, 1)] if rule.list_enabled else [(0, 0, 0, 0.85), (0.6, 0.5, 1, 1)]
-            self.hover_attr = (icon_path("close-circle.png"), 'DEMOTE', (1, 0.5, 0.65, 1)) if rule.list_enabled else (icon_path("promote.png"), 'PROMOTE', (0.3, 1, 0.6, 1))
-        elif utility.screen_manager.current_screen.current_list == "bans":
-            if rule.rule_type == "ip":
-                self.color_id = self.button.color_id = [(0, 0, 0, 0.85), (1, 0.45, 0.85, 1)] if rule.list_enabled else [(0, 0, 0, 0.85), (0.4, 0.8, 1, 1)]
-            else:
-                self.color_id = self.button.color_id = [(0, 0, 0, 0.85), (1, 0.5, 0.65, 1)] if rule.list_enabled else [(0, 0, 0, 0.85), (0.3, 1, 0.6, 1)]
-            if rule.list_enabled:
-                self.hover_attr = (icon_path("lock-open.png"), 'PARDON', (0.3, 1, 0.6, 1))
-            else:
-                self.hover_attr = (icon_path("close-circle.png"), 'BAN' if rule.rule_type == 'player' else 'REMOVE', (1, 0.5, 0.65, 1))
-        elif utility.screen_manager.current_screen.current_list == "wl":
-            if utility.screen_manager.current_screen.acl_object._server['whitelist']:
-                self.color_id = self.button.color_id = [(0, 0, 0, 0.85), (0.3, 1, 0.6, 1)] if rule.list_enabled else [(0, 0, 0, 0.85), (1, 0.5, 0.65, 1)]
-            else:
-                self.color_id = self.button.color_id = [(0, 0, 0, 0.85), (0.3, 1, 0.6, 1)] if rule.list_enabled else [(0, 0, 0, 0.85), (0.7, 0.7, 0.7, 1)]
-            self.hover_attr = (icon_path("close-circle.png"), 'RESTRICT', (1, 0.5, 0.65, 1)) if rule.list_enabled else (icon_path("checkmark-circle-sharp.png"), 'PERMIT', (0.3, 1, 0.6, 1))
-
-        new_color_id = (self.color_id[1][0], self.color_id[1][1], self.color_id[1][2], 1 if rule.list_enabled else 0.95)
-        self.color_id[1] = new_color_id
-        self.button.color_id[1] = new_color_id
-        self.text.color = constants.brighten_color(new_color_id, 0.2)
-        self.icon.color = constants.brighten_color(new_color_id, 0.2)
-        self.icon.source = icon_path("earth-sharp.png")
-        self.button.background_color = new_color_id
-
-
-        # Change scope attributes
-        if rule.rule_scope == "global":
-            self.text.pos_hint["center_x"] = 0.5 + round(len(self.text.text) * 0.01, 2)
-            self.icon.opacity = 1
-            self.icon.color = self.global_icon_color
-            if rule.rule_type == "player" or "/" in rule.rule:
-                self.text.font_size = sp(19 - (0 if len(self.text.text) < 11 else (len(self.text.text) // 5)))
-
-        else:
-            self.text.pos_hint["center_x"] = 0.5
-            self.icon.opacity = 0
-            self.icon.color = self.color_id[1]
-            if rule.rule_type == "player" or "/" in rule.rule:
-                self.text.font_size = sp(19 - (0 if len(self.text.text) < 11 else (len(self.text.text) // 7)))
-
-        if rule.rule_type == "ip" and "/" not in rule.rule:
-            self.text.font_size = sp(19)
-
-        self.original_font_size = self.text.font_size
-
-
-    def highlight(self, original_color, original_text_color, original_hover_color):
-        self.button.background_color = constants.brighten_color(original_hover_color, -0.15)
-        self.text.color = constants.brighten_color(original_hover_color, -0.15)
-        Animation(background_color=original_color, duration=1).start(self.button)
-        Animation(color=original_text_color, duration=1).start(self.text)
-
-
-    def __init__(self, name='', icon_name=None, width=None, icon_offset=None, auto_adjust_icon=False, **kwargs):
-        super().__init__(**kwargs)
-
-        position = (0.5, 0.5)
-        self.size_hint_max_y = dp(60)
-        self.rule = None
-        self.id = 'rule_button'
-        self.enabled = True
-        self.original_font_size = None
-        self.action_text = ''
-
-
-        # Hover button object
-        self.button = HoverButton()
-
-        def on_enter(*args):
-
-            Animation.cancel_all(self.button)
-            Animation.cancel_all(self.text)
-            Animation.cancel_all(self.icon)
-
-            if not self.button.ignore_hover:
-                self.text.font_size = sp(18)
-
-                if self.rule.rule_scope == "global":
-                    self.icon.source = icon_path("earth-strike.png")
-                    self.text.text = translate("LOCALIZE")
-                    self.action_text = "LOCALIZE"
-                    self.button.background_color = self.global_icon_color
-
-                else:
-                    self.icon.source = self.hover_attr[0]
-                    self.text.text = "   " + translate(self.hover_attr[1])
-                    self.action_text = self.hover_attr[1]
-                    self.button.background_color = self.hover_attr[2]
-                    Animation(opacity=1, duration=0.05).start(self.icon)
-
-                animate_button(self.button, image=os.path.join(paths.ui_assets, f'{self.button.id}_hover.png'), color=self.button.color_id[0], hover_action=True)
-
-        def on_leave(*args):
-
-            Animation.cancel_all(self.button)
-            Animation.cancel_all(self.text)
-
-            if not self.button.ignore_hover:
-                self.text.font_size = self.original_font_size
-                self.text.text = self.rule.rule.replace("!w", "")
-                new_color_id = (self.color_id[1][0], self.color_id[1][1], self.color_id[1][2], 1 if self.rule.list_enabled else 0.95)
-                animate_button(self.button, image=os.path.join(paths.ui_assets, f'{self.button.id}.png'), color=constants.brighten_color(self.button.color_id[1], 0.2), hover_action=False, _new_color=new_color_id)
-                Animation.cancel_all(self.icon)
-                if self.rule.rule_scope == "global":
-                    Animation(color=self.global_icon_color, duration=0.1).start(self.icon)
-                    self.icon.source = icon_path("earth-sharp.png")
-                else:
-                    Animation(opacity=0, duration=0.05).start(self.icon)
-
-        def click_func(button_pressed=None, *args):
-
-            if not button_pressed or not isinstance(button_pressed, str):
-                button_pressed = self.button.button_pressed.lower().strip()
-
-            button_text = self.action_text.lower().strip()
-            current_list = utility.screen_manager.current_screen.current_list.lower().strip()
-            acl_object = utility.screen_manager.current_screen.acl_object
-            original_name = self.rule.rule
-            filtered_name = self.rule.rule.replace("!w", "").replace("!g", "").strip()
-            original_hover_attr = self.hover_attr
-            new_scope = self.rule.rule_scope
-            banner_text = ""
-            reload_page = False
-            localize = False
-
-
-            # Left click on button (Local)
-            if button_pressed == "left" and self.rule.rule_scope == "local":
-
-                # Modify 'ops' list
-                if current_list == "ops" and button_text in ['promote', 'demote']:
-                    acl_object.op_player(self.rule.rule, remove=(button_text == "demote"))
-                    banner_text = f"'${filtered_name}$' was {'demoted' if (button_text == 'demote') else 'promoted'}"
-                    reload_page = True
-
-                # Modify 'bans' list
-                elif current_list == "bans" and button_text in ['ban', 'pardon', 'remove']:
-                    acl_object.ban_player(self.rule.rule, remove=(button_text == "pardon") or ("!w" in self.rule.rule))
-
-                    if button_text == "pardon":
-                        try:
-                            ip_addr = acl.get_uuid(self.rule.rule)['latest-ip'].split(":")[0].strip()
-                        except KeyError:
-                            ip_addr = ""
-
-                        if ip_addr:
-                            # Whitelist IP if it's still in the rule list
-                            if ip_addr in acl.gen_iplist(acl_object.rules['subnets']):
-                                acl_object.ban_player(f"!w{ip_addr}", remove=False)
-
-                    banner_text = f"'${filtered_name}$' was removed" if button_text == 'remove' else f"'{filtered_name}' is {'pardoned' if (button_text == 'pardon') else 'banned'}"
-                    reload_page = True
-
-                # Modify 'wl' list
-                elif current_list == "wl" and button_text in ['permit', 'restrict']:
-                    acl_object.whitelist_player(self.rule.rule, remove=(button_text == "restrict"))
-                    banner_text = f"'${filtered_name}$' is {'restricted' if (button_text == 'restrict') else 'permitted'}"
-                    reload_page = True
-
-
-            # Left click on button (global)
-            elif button_pressed == "left" and self.rule.rule_scope == "global" and button_text == "localize":
-                acl_object.add_global_rule(self.rule.rule, current_list, remove=True)
-                original_hover_attr = (
-                    icon_path("earth-strike.png"),
-                    original_hover_attr[1],
-                    self.color_id[1]
-                )
-                banner_text = f"'${filtered_name}$' rule is now locally applied"
-                localize = True
-                new_scope = "local"
-                reload_page = True
-
-
-            # Middle click on button
-            elif button_pressed == "middle":
-                acl_object.add_global_rule(self.rule.rule, current_list, remove=(button_text == 'localize'))
-                original_hover_attr = (
-                    icon_path(f"earth-{'strike' if (self.rule.rule_scope == 'global') else 'sharp'}.png"),
-                    original_hover_attr[1],
-                    self.global_icon_color if (self.rule.rule_scope == 'local') else self.color_id[1]
-                )
-                banner_text = f"'${filtered_name}$' is now {'local' if (self.rule.rule_scope == 'global') else 'global'}ly applied"
-                localize = self.rule.rule_scope == "global"
-                new_scope = "local" if localize else "global"
-                reload_page = True
-
-
-
-            if reload_page:
-                # print(button_text, button_pressed, current_list)
-
-                # If rule localized, enable on current list
-                if localize:
-                    if current_list == "ops":
-                        acl_object.op_player(self.rule.rule)
-                    elif current_list == "bans":
-                        acl_object.ban_player(self.rule.rule)
-                    elif current_list == "wl":
-                        acl_object.whitelist_player(self.rule.rule)
-
-
-                self.button.on_leave()
-                self.change_properties(self.rule)
-                self.button.state = "normal"
-
-                Animation.cancel_all(self.button)
-                Animation.cancel_all(self.text)
-                Animation.cancel_all(self.icon)
-
-                utility.screen_manager.current_screen.update_list(current_list, reload_children=False)
-
-                Clock.schedule_once(
-                    functools.partial(
-                        utility.screen_manager.current_screen.show_banner,
-                        original_hover_attr[2],
-                        banner_text,
-                        original_hover_attr[0],
-                        2,
-                        {"center_x": 0.5, "center_y": 0.965}
-                    ), 0
-                )
-
-                def trigger_highlight(*args):
-                    for rule_button in utility.screen_manager.current_screen.scroll_layout.children:
-
-                        if rule_button.rule.rule == original_name:
-                            rule_button.highlight(rule_button.button.background_color, rule_button.text.color, original_hover_attr[2])
-
-                        else:
-                            rule_button.button.on_leave()
-                            Animation.cancel_all(rule_button.button)
-                            Animation.cancel_all(rule_button.text)
-                            Animation.cancel_all(rule_button.icon)
-                            rule_button.change_properties(rule_button.rule)
-
-                        rule_button.button.ignore_hover = False
-
-                Clock.schedule_once(trigger_highlight, 0)
-
-            # Update display rule regardless of button pressed
-            utility.screen_manager.current_screen.update_user_panel(self.rule.rule, new_scope)
-
-
-
-        self.button.on_enter = on_enter
-        self.button.on_leave = on_leave
-        self.button.bind(on_press=click_func)
-        self.button.id = 'rule_button'
-        self.color_id = self.button.color_id = [(0.03, 0.03, 0.03, 1), (1, 1, 1, 1)] # [(0.05, 0.05, 0.1, 1), (0.6, 0.6, 1, 1)]
-        self.hover_attr = (icon_path('close-circle.png'), 'hover text', (1, 1, 1, 1)) # Icon, Text, Hover color
-        self.global_icon_color = (0.953, 0.929, 0.38, 1)
-
-        self.button.size_hint = (None, None)
-        self.button.size = (dp(190 if not width else width), self.size_hint_max_y)
-        self.button.pos_hint = {"center_x": position[0], "center_y": position[1]}
-        self.button.border = (-3, -3, -3, -3)
-        self.button.background_normal = os.path.join(paths.ui_assets, 'rule_button.png')
-        self.button.background_down = os.path.join(paths.ui_assets, 'rule_button_click.png')
-        self.button.always_release = True
-
-        self.text = Label()
-        self.text.__translate__ = False
-        self.text.id = 'text'
-        self.text.size_hint = (None, None)
-        self.text.pos_hint = {"center_x": position[0], "center_y": position[1]}
-        self.text.text = name
-        self.text.font_size = sp(19)
-        self.text.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["bold"]}.ttf')
-        self.text.color = self.color_id[1]
-
-
-        # Button click behavior
-        self.button.on_release = functools.partial(button_action, name, self.button)
-        self.add_widget(self.button)
-
-
-        # Icon thingy
-        self.icon = Image()
-        self.icon.id = 'icon'
-        self.icon.source = icon_path("earth-sharp.png") if not icon_name else icon_path(icon_name)
-        self.icon.size = (25, 25)
-        self.icon.size_hint = (None, None)
-        self.icon.opacity = 0
-        self.icon.color = self.color_id[1]
-        self.icon.pos_hint = {"center_x": -0.2, "center_y": 0.5}
-
-
-        self.add_widget(self.icon)
-        self.add_widget(self.text)
-
-class AclRulePanel(RelativeLayout):
-
-    def __init__(self, **kw):
-        super().__init__(**kw)
-
-        class HeaderLabel(Label):
-            def __init__(self, **kwargs):
-                super().__init__(**kwargs)
-                self.size_hint = (None, None)
-                self.markup = True
-                self.font_size = sp(22)
-                self.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["medium"]}.ttf')
-                self.color = (0.6, 0.6, 1, 1)
-
-        class ParagraphLabel(Label, HoverBehavior):
-
-            def on_mouse_pos(self, *args):
-
-                if "AclScreen" in utility.screen_manager.current_screen.name:
-
-                    try:
-                        super().on_mouse_pos(*args)
-                    except:
-                        pass
-
-                    if self.text.count(".") > 3 and "IP" in self.text:
-                        rel_y = args[1][1] - utility.screen_manager.current_screen.user_panel.y
-                        if self.hovered and rel_y < 190:
-                            self.on_leave()
-                            self.hovered = False
-
-
-            # Hover stuffies
-            def on_enter(self, *args):
-
-                # Change size of IP text
-                if self.text.count(".") > 3 and "IP" in self.text:
-                    rel_y = self.border_point[1] - utility.screen_manager.current_screen.user_panel.y
-                    if rel_y < 190:
-                        self.hovered = False
-                        return None
-
-                if self.copyable:
-                    self.outline_width = 0
-                    self.outline_color = constants.brighten_color(self.color, 0.05)
-                    Animation(outline_width=1, duration=0.03).start(self)
-
-
-            def on_leave(self, *args):
-
-                if self.copyable:
-                    Animation.stop_all(self)
-                    self.outline_width = 0
-
-
-            # Normal stuffies
-            def on_ref_press(self, *args):
-                if not self.disabled:
-
-                    Clock.schedule_once(
-                        functools.partial(
-                            utility.screen_manager.current_screen.show_banner,
-                            (0.85, 0.65, 1, 1),
-                            "Copied text to clipboard",
-                            "link-sharp.png",
-                            2,
-                            {"center_x": 0.5, "center_y": 0.965}
-                        ), 0
-                    )
-
-                    Clipboard.copy(re.sub(r"\[.*?\]","",self.text))
-
-
-            def ref_text(self, *args):
-
-                self.copyable = not ((translate("unknown") in self.text.lower()) or (translate("online") in self.text.lower()) or (translate("access") in self.text.lower()))
-
-                if '[ref=' not in self.text and '[/ref]' not in self.text and self.copyable:
-                    self.text = f'[ref=none]{self.text}[/ref]'
-                elif '[/ref]' in self.text:
-                    self.text = self.text.replace("[/ref]","") + "[/ref]"
-
-                self.texture_update()
-                self.size = self.texture_size
-
-                if self.text.count(".") > 3 and "IP" in self.text:
-                    self.width = self.texture_size[0] / 1.5
-
-
-            def __init__(self, **kwargs):
-                super().__init__(**kwargs)
-                self.size_hint = (None, None)
-                self.markup = True
-                self.font_size = sp(18)
-                self.copyable = True
-                self.font_name = os.path.join(paths.ui_assets, 'fonts', f'{constants.fonts["regular"]}.ttf')
-                self.default_color = (0.6, 0.6, 1, 1)
-                self.color = self.default_color
-                self.bind(text=self.ref_text)
-
-
-        self.color_dict = {
-            "blue":    "#70E6FF",
-            "red":     "#FF8793",
-            "green":   "#4CFF99",
-            "white":   "#FFFFFF",
-            "gray":    "#A0A0A0",
-            "yellow":  "#F3ED61",
-            "purple":  "#A699FF"
-        }
-
-
-        self.displayed_type = ""
-        self.displayed_scope = "local"
-
-        # User panel (I'm sorry in advance)
-        # <editor-fold desc="User Panel SUFFERING">
-        self.pos_hint = {"center_y": 0.42}
-        self.size_hint_max = (500, 600)
-
-        # Background image
-        self.background = Image()
-        self.background.id = 'background'
-        self.background.source = os.path.join(paths.ui_assets, 'user_panel.png')
-        self.background.color = (0.65, 0.6, 1, 1)
-        self.background.opacity = 0
-
-        # Label when no rule is displayed
-        self.blank_label = Label()
-        self.blank_label.id = 'blank_label'
-        self.blank_label.text = "Right-click a rule to view"
-        self.blank_label.text_size[0] = self.size_hint_max[0] * 0.7
-        self.blank_label.halign = "center"
-        self.blank_label.font_name = os.path.join(paths.ui_assets, 'fonts', constants.fonts['italic'])
-        self.blank_label.pos_hint = {"center_x": 0.5, "center_y": 0.5}
-        self.blank_label.font_size = sp(26)
-        self.blank_label.color = (0.6, 0.6, 1, 0.4)
-        self.blank_label.opacity = 0
-
-        # Drop down options button
-        self.options = DropButton('OPTIONS', (0.5, 0.192), options_list=['operators', 'bans', 'whitelist'], input_name='ServerAclOptionsInput', x_offset=-33, facing='center', custom_func=self.modify_rule, change_text=False)
-        self.add_widget(self.options)
-
-
-        # Player Layout
-        # <editor-fold desc="Player Layout Widgets">
-        self.player_layout = RelativeLayout()
-        self.player_layout.id = 'player_layout'
-
-        self.player_layout.header_icon = Image()
-        self.player_layout.header_icon.source = icon_path('person-circle-sharp.png')
-        self.player_layout.header_icon.allow_stretch = True
-        self.player_layout.header_icon.size_hint_max = (36, 36)
-        self.player_layout.header_icon.pos_hint = {"center_x": 0.5, "center_y": 0.808}
-        self.player_layout.add_widget(self.player_layout.header_icon)
-
-        # Make this copyable text
-        self.player_layout.name_label = HeaderLabel()
-        self.player_layout.name_label.__translate__ = False
-        self.player_layout.name_label.pos_hint = {"center_x": 0.54, "center_y": 0.81}
-        self.player_layout.name_label.font_size = sp(25)
-        self.player_layout.add_widget(self.player_layout.name_label)
-
-        self.player_layout.online_icon = Image()
-        self.player_layout.online_icon.source = icon_path('radio-button-off-sharp.png')
-        self.player_layout.online_icon.allow_stretch = True
-        self.player_layout.online_icon.size_hint_max = (15, 15)
-        self.player_layout.online_icon.pos_hint = {"center_x": 0.5, "center_y": 0.752}
-        self.player_layout.add_widget(self.player_layout.online_icon)
-
-        self.player_layout.online_label = ParagraphLabel()
-        self.player_layout.online_label.font_size = sp(19)
-        self.player_layout.online_label.pos_hint = {"center_x": 0.523, "center_y": 0.755}
-        self.player_layout.add_widget(self.player_layout.online_label)
-
-        self.player_layout.uuid_header = HeaderLabel()
-        self.player_layout.uuid_header.__translate__ = False
-        self.player_layout.uuid_header.pos_hint = {"center_x": 0.5, "center_y": 0.66}
-        self.player_layout.add_widget(self.player_layout.uuid_header)
-
-        # Make this copyable text
-        self.player_layout.uuid_label = ParagraphLabel()
-        self.player_layout.uuid_label.__translate__ = False
-        self.player_layout.uuid_label.pos_hint = {"center_x": 0.5, "center_y": 0.619}
-        self.player_layout.add_widget(self.player_layout.uuid_label)
-
-        self.player_layout.ip_header = HeaderLabel()
-        self.player_layout.ip_header.__translate__ = False
-        self.player_layout.ip_header.pos_hint = {"center_x": 0.28, "center_y": 0.54}
-        self.player_layout.add_widget(self.player_layout.ip_header)
-
-        # Make this copyable text
-        self.player_layout.ip_label = ParagraphLabel()
-        self.player_layout.ip_label.__translate__ = False
-        self.player_layout.ip_label.font_size = sp(20)
-        self.player_layout.ip_label.pos_hint = {"center_x": 0.28, "center_y": 0.499}
-        self.player_layout.add_widget(self.player_layout.ip_label)
-
-        self.player_layout.geo_header = HeaderLabel()
-        self.player_layout.geo_header.__translate__ = False
-        self.player_layout.geo_header.pos_hint = {"center_x": 0.7, "center_y": 0.54}
-        self.player_layout.add_widget(self.player_layout.geo_header)
-
-        self.player_layout.geo_label = ParagraphLabel()
-        self.player_layout.geo_label.__translate__ = False
-        self.player_layout.geo_label.halign = "center"
-        self.player_layout.geo_label.pos_hint = {"center_x": 0.7, "center_y": 0.499}
-        self.player_layout.add_widget(self.player_layout.geo_label)
-
-        self.player_layout.access_header = HeaderLabel()
-        self.player_layout.access_header.pos_hint = {"center_x": 0.5, "center_y": 0.4}
-        self.player_layout.access_header.font_size = sp(20)
-        self.player_layout.add_widget(self.player_layout.access_header)
-
-        self.player_layout.access_label = ParagraphLabel()
-        self.player_layout.access_label.halign = "left"
-        self.player_layout.access_label.valign = "top"
-        self.player_layout.access_label.text_size = (250, 300)
-        self.player_layout.access_label.font_size = sp(19)
-        self.player_layout.access_label.line_height = sp(1.4)
-        self.player_layout.access_label.pos_hint = {"center_x": 0.465, "center_y": 0.11}
-        self.player_layout.add_widget(self.player_layout.access_label)
-
-        self.player_layout.access_line_1 = Image()
-        self.player_layout.access_line_1.size_hint_max = (35, 35)
-        self.player_layout.access_line_1.pos_hint = {"center_x": 0.165, "center_y": 0.29}
-        self.player_layout.access_line_1.allow_stretch = True
-        self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_active.png")
-        self.player_layout.add_widget(self.player_layout.access_line_1)
-
-        self.player_layout.access_line_2 = Image()
-        self.player_layout.access_line_2.size_hint_max = (35, 35)
-        self.player_layout.access_line_2.pos_hint = {"center_x": 0.165, "center_y": 0.24}
-        self.player_layout.access_line_2.allow_stretch = True
-        self.player_layout.access_line_2.source = os.path.join(paths.ui_assets, "access_active.png")
-        self.player_layout.add_widget(self.player_layout.access_line_2)
-
-        self.player_layout.access_line_3 = Image()
-        self.player_layout.access_line_3.size_hint_max = (35, 35)
-        self.player_layout.access_line_3.pos_hint = {"center_x": 0.165, "center_y": 0.19}
-        self.player_layout.access_line_3.allow_stretch = True
-        self.player_layout.access_line_3.source = os.path.join(paths.ui_assets, "access_active.png")
-        self.player_layout.add_widget(self.player_layout.access_line_3)
-
-        self.player_layout.access_icon = Image()
-        self.player_layout.access_icon.size_hint_max = (30, 30)
-        self.player_layout.access_icon.pos_hint = {"center_x": 0.165, "center_y": 0.338}
-        self.player_layout.add_widget(self.player_layout.access_icon)
-        # </editor-fold>
-
-
-        # IP Layout
-        # <editor-fold desc="IP Layout Widgets">
-        self.ip_layout = RelativeLayout()
-        self.ip_layout.id = 'ip_layout'
-
-        self.ip_layout.header_icon = Image()
-        self.ip_layout.header_icon.source = icon_path('ethernet.png')
-        self.ip_layout.header_icon.allow_stretch = True
-        self.ip_layout.header_icon.size_hint_max = (36, 36)
-        self.ip_layout.header_icon.pos_hint = {"center_x": 0.5, "center_y": 0.808}
-        self.ip_layout.add_widget(self.ip_layout.header_icon)
-
-        # Make this copyable text
-        self.ip_layout.name_label = HeaderLabel()
-        self.ip_layout.name_label.__translate__ = False
-        self.ip_layout.name_label.pos_hint = {"center_x": 0.54, "center_y": 0.81}
-        self.ip_layout.name_label.font_size = sp(25)
-        self.ip_layout.add_widget(self.ip_layout.name_label)
-
-        self.ip_layout.type_header = HeaderLabel()
-        self.ip_layout.type_header.pos_hint = {"center_x": 0.28, "center_y": 0.64}
-        self.ip_layout.add_widget(self.ip_layout.type_header)
-
-        # Make ip copyable text
-        self.ip_layout.type_label = ParagraphLabel()
-        self.ip_layout.type_label.__translate__ = False
-        self.ip_layout.type_label.font_size = sp(20)
-        self.ip_layout.type_label.pos_hint = {"center_x": 0.28, "center_y": 0.598}
-        self.ip_layout.add_widget(self.ip_layout.type_label)
-
-        self.ip_layout.affected_header = HeaderLabel()
-        self.ip_layout.affected_header = HeaderLabel()
-        self.ip_layout.affected_header.pos_hint = {"center_x": 0.7, "center_y": 0.64}
-        self.ip_layout.add_widget(self.ip_layout.affected_header)
-
-        # Make ip copyable text
-        self.ip_layout.affected_label = ParagraphLabel()
-        self.ip_layout.affected_label.__translate__ = False
-        self.ip_layout.affected_label.halign = "center"
-        self.ip_layout.affected_label.font_size = sp(20)
-        self.ip_layout.affected_label.pos_hint = {"center_x": 0.7, "center_y": 0.598}
-        self.ip_layout.add_widget(self.ip_layout.affected_label)
-
-        self.ip_layout.network_header = HeaderLabel()
-        self.ip_layout.network_header.pos_hint = {"center_x": 0.5, "center_y": 0.458}
-        self.ip_layout.add_widget(self.ip_layout.network_header)
-
-        # Make IP copyable text
-        self.ip_layout.network_label = ParagraphLabel()
-        self.ip_layout.network_label.__translate__ = False
-        self.ip_layout.network_label.halign = "center"
-        self.ip_layout.network_label.valign = "top"
-        self.ip_layout.network_label.text_size = (400, 150)
-        self.ip_layout.network_label.font_size = sp(20)
-        self.ip_layout.network_label.line_height = sp(1.4)
-        self.ip_layout.network_label.pos_hint = {"center_x": 0.5, "center_y": 0.3}
-        self.ip_layout.add_widget(self.ip_layout.network_label)
-        # </editor-fold>
-
-
-        for widget in self.player_layout.children:
-            utility.hide_widget(widget, True)
-        for widget in self.ip_layout.children:
-            utility.hide_widget(widget, True)
-
-        self.add_widget(self.background)
-        self.add_widget(self.blank_label)
-        self.add_widget(self.player_layout)
-        self.add_widget(self.ip_layout)
-        # </editor-fold>
-
-
-    # Actually updates data in panel based off of rule
-    def update_panel(self, displayed_rule: acl.AclRule, rule_scope: str):
-
-        self.displayed_scope = rule_scope
-        filtered_name = displayed_rule.rule.replace("!g", "").replace("!w", "")
-        panel_options = []
-
-        # Player layout ------------------------------------------------------------------------------------------------
-        if displayed_rule.rule_type == "player":
-
-            # Effective access colors
-            if displayed_rule.display_data['effective_access'] == translate("Operator access"):
-                widget_color = self.color_dict['blue']
-                self.player_layout.access_icon.source = icon_path('promote.png')
-            elif displayed_rule.display_data['effective_access'] == translate("No access"):
-                widget_color = self.color_dict['red']
-                self.player_layout.access_icon.source = icon_path('close-circle-outline.png')
-            else:
-                widget_color = self.color_dict['purple']
-                self.player_layout.access_icon.source = icon_path('chevron-up-circle-sharp.png')
-
-
-            if self.displayed_type != "player":
-                for widget in self.player_layout.children:
-                    utility.hide_widget(widget, False)
-                for widget in self.ip_layout.children:
-                    utility.hide_widget(widget, True)
-
-
-            # Change panel attributes ----------------------------------------------------------------------------------
-
-
-            # Change name in header
-            self.player_layout.name_label.text = filtered_name
-            self.player_layout.name_label.texture_update()
-            texture_size = 0.001 * self.player_layout.name_label.texture_size[0]
-            self.player_layout.header_icon.pos_hint = {"center_x": 0.485 - texture_size, "center_y": 0.808}
-
-            # Change icon in header
-            # self.player_layout.header_icon.source = os.path.join(paths.gui_assets, 'steve.png')
-            # def update_source(*a):
-            #     source = manager.get_player_head(filtered_name)
-            #     def main_thread(*b):
-            #         if self.player_layout.name_label.text == filtered_name:
-            #             self.player_layout.header_icon.source = source
-            #     Clock.schedule_once(main_thread, 0)
-            # dTimer(0, update_source).start()
-
-
-            # Online status
-            if constants.server_manager.current_server and constants.server_manager.current_server.user_online(displayed_rule.rule):
-                self.player_layout.online_label.color = self.color_dict['green']
-                self.player_layout.online_icon.color = self.color_dict['green']
-                self.player_layout.online_label.text = "Currently online"
-                self.player_layout.online_icon.source = icon_path('radio-button-on-sharp.png')
-
-            else:
-                self.player_layout.online_label.color = self.color_dict['gray']
-                self.player_layout.online_icon.color = self.color_dict['gray']
-                try:
-                    last_login = (dt.now() - dt.strptime(displayed_rule.extra_data['latest-login'], '%Y-%m-%d %H:%M:%S'))
-                    d = {"days": last_login.days}
-                    d["years"], rem = divmod(last_login.days, 365)
-                    d["months"], rem = divmod(last_login.days, 30)
-                    d["hours"], rem = divmod(last_login.seconds, 3600)
-                    d["minutes"], d["seconds"] = divmod(rem, 60)
-
-                    if d['years'] > 0:
-                        time_formatted = (f"{d['years']} {translate('year' + ('s' if d['years'] > 1 else ''))} " if d['years'] > 0 else "")
-                    elif d['months'] > 0:
-                        time_formatted = (f"{d['months']} {translate('month' + ('s' if d['months'] > 1 else ''))} " if d['months'] > 0 else "")
-                    else:
-                        time_formatted = (f"{d['days']}d " if d['days'] > 0 else "") + (f"{d['hours']}h " if d['hours'] > 0 else "") + (f"{d['minutes']}m " if d['minutes'] > 0 and d['days'] == 0 else "")
-
-                    if not time_formatted:
-                        time_formatted = 'seconds '
-
-                    self.player_layout.online_label.text = f"Last online ${time_formatted}$ago"
-
-                except ValueError:
-                    self.player_layout.online_label.text = "Last online unknown"
-
-                self.player_layout.online_icon.source = icon_path('radio-button-off-sharp.png')
-
-            self.player_layout.online_label.texture_update()
-            texture_size = 0.001 * self.player_layout.online_label.texture_size[0]
-            self.player_layout.online_icon.pos_hint = {"center_x": 0.492 - texture_size, "center_y": 0.752}
-
-
-            # Change UUID
-            self.player_layout.uuid_label.text = displayed_rule.extra_data['uuid']
-
-
-            # Change last IP
-            if displayed_rule.extra_data['latest-ip'].startswith("127."):
-                self.player_layout.ip_label.text = displayed_rule.extra_data['latest-ip']
-            else:
-                self.player_layout.ip_label.text = displayed_rule.extra_data['latest-ip'].split(":")[0]
-
-
-            # Change location
-            if " - " in displayed_rule.extra_data['ip-geo']:
-                self.player_layout.geo_label.text = displayed_rule.extra_data['ip-geo'].replace(" - ", "\n")
-                self.player_layout.geo_label.pos_hint = {"center_x": 0.7, "center_y": 0.48}
-            else:
-                self.player_layout.geo_label.text = displayed_rule.extra_data['ip-geo']
-                self.player_layout.geo_label.pos_hint = {"center_x": 0.7, "center_y": 0.499}
-            self.player_layout.geo_label.font_size = sp(20) if len(displayed_rule.extra_data['ip-geo']) < 15 else sp(18)
-
-
-            # Change effective access
-            very_bold_font = os.path.join(paths.ui_assets, 'fonts', constants.fonts["bold"])
-            final_text = f"[color={widget_color}][font={very_bold_font}][size={round(sp(21))}]{displayed_rule.display_data['effective_access']}[/size][/font][/color]"
-            banned = False
-
-            # Display ban data
-            if displayed_rule.display_data['ip_ban'] and displayed_rule.display_data['ban']:
-                final_text += f"\n[color={self.color_dict['red']}]{translate('Banned IP & user')}[/color]"
-                banned = True
-            elif displayed_rule.display_data['ip_ban']:
-                final_text += f"\n[color={self.color_dict['red']}]{translate('Banned IP')}[/color]"
-                banned = True
-            elif displayed_rule.display_data['ban']:
-                final_text += f"\n[color={self.color_dict['red']}]{translate('Banned user')}[/color]"
-                banned = True
-            else:
-                final_text += "\n"
-
-            # Display OP data
-            final_text += (f"\n[color={self.color_dict['blue']}]{translate('Operator')}[/color]" if displayed_rule.display_data['op'] else "\n")
-
-            # Whitelist data
-            if utility.screen_manager.current_screen.acl_object._server['whitelist']:
-                self.player_layout.access_line_3.opacity = 1
-                if constants.app_config.locale == 'en':
-                    final_text += ("\n[color=" + (f"{self.color_dict['green']}]Whitelisted" if displayed_rule.display_data['wl'] else f"{self.color_dict['red']}]Not whitelisted") + "[/color]")
-                else:
-                    final_text += ("\n[color=" + (f"{self.color_dict['green']}]{translate('Allowed')}" if displayed_rule.display_data['wl'] else f"{self.color_dict['red']}]{translate('Denied')}") + "[/color]")
-            else:
-                self.player_layout.access_line_3.opacity = 0
-                self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_active.png")
-                self.player_layout.access_line_2.source = os.path.join(paths.ui_assets, "access_active.png")
-                self.player_layout.access_line_3.source = os.path.join(paths.ui_assets, "access_active.png")
-
-            self.player_layout.access_label.text = final_text
-
-
-            # Adjust graphic data for access
-            if utility.screen_manager.current_screen.acl_object._server['whitelist']:
-                if displayed_rule.display_data['wl']:
-                    self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_active.png")
-                    self.player_layout.access_line_2.source = os.path.join(paths.ui_assets, "access_active.png")
-                    self.player_layout.access_line_3.source = os.path.join(paths.ui_assets, "access_active.png")
-                else:
-                    self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_inactive.png")
-                    self.player_layout.access_line_2.source = os.path.join(paths.ui_assets, "access_inactive.png")
-                    self.player_layout.access_line_3.source = os.path.join(paths.ui_assets, "access_inactive.png")
-
-                    if displayed_rule.display_data['op']:
-                        self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_active.png")
-                        self.player_layout.access_line_2.source = os.path.join(paths.ui_assets, "access_active.png")
-
-            if banned:
-                self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_inactive.png")
-            elif displayed_rule.display_data['wl']:
-                self.player_layout.access_line_1.source = os.path.join(paths.ui_assets, "access_active.png")
-
-
-            # Set header names
-            self.player_layout.uuid_header.text = "UUID"
-            self.player_layout.ip_header.text = "IP"
-            self.player_layout.geo_header.text = "Location"
-            self.player_layout.access_header.text = f"Access to '${utility.screen_manager.current_screen.acl_object._server['name']}$':"
-
-
-
-            # Change colors based on rule access attributes ------------------------------------------------------------
-            self.background.color = widget_color
-            for widget in self.player_layout.children:
-
-                if widget is self.player_layout.online_label or widget is self.player_layout.online_icon:
-                    continue
-
-                if widget.__class__.__name__ in ['Image', 'HeaderLabel']:
-                    if widget in [self.player_layout.header_icon, self.player_layout.name_label]:
-                        widget.color = constants.brighten_color(widget_color, 0.12)
-                    elif widget in [self.player_layout.access_icon, self.player_layout.access_line_1, self.player_layout.access_line_2, self.player_layout.access_line_3]:
-                        widget.color = constants.brighten_color(widget_color, 0.2)
-                    else:
-                        widget.color = constants.brighten_color(widget_color, 0.34)
-
-                elif widget.__class__.__name__ == "ParagraphLabel":
-                    widget.color = self.color_dict['gray'] if translate("unknown") in widget.text.lower() else widget.default_color
-
-
-
-            # Generate panel options -----------------------------------------------------------------------------------
-            if utility.screen_manager.current_screen.current_list == "ops":
-                panel_options.append("demote" if displayed_rule.display_data['op'] else "promote")
-
-            elif utility.screen_manager.current_screen.current_list == "bans":
-                if displayed_rule.display_data['ip_ban'] and displayed_rule.display_data['ban']:
-                    panel_options.append("pardon IP & user")
-                elif displayed_rule.display_data['ip_ban']:
-                    if not displayed_rule.display_data['ban']:
-                        panel_options.append("ban user")
-                    panel_options.append("pardon IP")
-                elif displayed_rule.display_data['ban']:
-                    panel_options.append("pardon user")
-                else:
-                    panel_options.append("ban user")
-
-                if displayed_rule.extra_data['latest-ip'] != "Unknown" and not displayed_rule.display_data['ip_ban']:
-                    panel_options.append("ban IP")
-
-            elif utility.screen_manager.current_screen.current_list == "wl":
-                panel_options.append("restrict" if displayed_rule.display_data['wl'] else "permit")
-
-            panel_options.append("localize rule" if rule_scope == "global" else "globalize rule")
-
-
-        # IP Layout ----------------------------------------------------------------------------------------------------
-        else:
-
-            # Effective access colors
-            if "whitelist" in displayed_rule.display_data['rule_info'].lower():
-                widget_color = self.color_dict['blue']
-            else:
-                widget_color = self.color_dict['red']
-
-            if not utility.screen_manager.current_screen.acl_object.rule_in_acl(displayed_rule.rule, 'subnets'):
-                displayed_rule.display_data['rule_info'] = "Unaffected " + displayed_rule.display_data['rule_info'].split(" ")[0]
-                displayed_rule.rule = displayed_rule.rule.replace("!w", "").replace("!g", "").strip()
-                utility.screen_manager.current_screen.displayed_rule = displayed_rule
-                widget_color = self.color_dict['purple']
-
-            if self.displayed_type != "ip":
-                for widget in self.player_layout.children:
-                    utility.hide_widget(widget, True)
-                for widget in self.ip_layout.children:
-                    utility.hide_widget(widget, False)
-
-
-
-            # Change panel attributes ----------------------------------------------------------------------------------
-
-
-            # Change IP address in header
-            self.ip_layout.name_label.text = filtered_name
-            self.ip_layout.name_label.texture_update()
-            texture_size = 0.001 * self.ip_layout.name_label.texture_size[0]
-            self.ip_layout.header_icon.pos_hint = {"center_x": 0.485 - texture_size, "center_y": 0.808}
-
-
-            # Change rule type
-            self.ip_layout.type_label.text = displayed_rule.display_data['rule_info']
-
-
-            # Change affected users
-            users = displayed_rule.display_data['affected_users']
-            if users == 0:
-                self.ip_layout.affected_label.text = f"0 {translate('users')}"
-                self.ip_layout.affected_label.color = self.color_dict['gray']
-            else:
-                self.ip_layout.affected_label.text = f"{users:,} {translate('user' + 's' if users > 1 else '')}"
-                self.ip_layout.affected_label.color = self.color_dict['green'] if "whitelist" in displayed_rule.display_data['rule_info'].lower() else self.color_dict['red']
-
-
-            # Change network info
-            ip_count = acl.count_subnet(displayed_rule.rule.replace("!w", "").replace("!g", "").strip())
-            ips = displayed_rule.display_data['ip_range'].split(" - ")
-            self.ip_layout.network_label.text = f"{ips[0]} [color=#696997]-[/color] {ips[1]}"
-            self.ip_layout.network_label.text += f"\n[color={self.color_dict['gray']}]{displayed_rule.display_data['subnet_mask']}"
-            self.ip_layout.network_label.text += f"  ({ip_count:,} IP{'s' if ip_count > 1 else ''})[/color]"
-
-
-
-            # Set header names
-            self.ip_layout.type_header.text = "Rule Type"
-            self.ip_layout.affected_header.text = "Affected"
-            self.ip_layout.network_header.text = "Network"
-
-
-            # Change colors based on rule access attributes ------------------------------------------------------------
-            self.background.color = widget_color
-            for widget in self.ip_layout.children:
-                if widget.__class__.__name__ in ['Image', 'HeaderLabel']:
-                    if widget in [self.ip_layout.header_icon, self.ip_layout.name_label]:
-                        widget.color = constants.brighten_color(widget_color, 0.12)
-                    else:
-                        widget.color = constants.brighten_color(widget_color, 0.34)
-
-
-            # Generate panel options -----------------------------------------------------------------------------------
-            if utility.screen_manager.current_screen.current_list == "bans":
-                after_text = 'subnet' if 'subnet' in displayed_rule.display_data['rule_info'].lower() else 'IP'
-                if "unaffected" in displayed_rule.display_data['rule_info'].lower():
-                    panel_options.append(f"ban {after_text}")
-                    panel_options.append(f"whitelist {after_text}")
-
-                else:
-                    if "whitelist" in displayed_rule.display_data['rule_info'].lower():
-                        panel_options.append("remove rule")
-                        panel_options.append(f"ban {after_text}")
-                    else:
-                        panel_options.append(f"pardon {after_text}")
-                        panel_options.append(f"whitelist {after_text}")
-
-                    panel_options.append("localize rule" if rule_scope == "global" else "globalize rule")
-
-        panel_options = panel_options if panel_options else ['no options']
-        self.options.change_options(panel_options)
-        self.displayed_type = displayed_rule.rule_type
-
-
-    # Changes rule attributes from options drop-down
-    def modify_rule(self, option: str):
-
-        current_list = utility.screen_manager.current_screen.current_list.lower().strip()
-        acl_object = utility.screen_manager.current_screen.acl_object
-        original_name = acl_object.displayed_rule.rule
-        filtered_name = acl_object.displayed_rule.rule.replace("!w", "").replace("!g", "").strip()
-        new_name = original_name
-        new_scope = self.displayed_scope
-        banner_text = ""
-        hover_attr = None
-        reload_page = False
-        localize = False
-
-        try:
-            ip_addr = acl_object.displayed_rule.extra_data['latest-ip'].split(":")[0].strip()
-        except KeyError:
-            ip_addr = ""
-
-
-        # Global options
-        if "local" in option or "global" in option:
-            if "localize" in option:
-                acl_object.add_global_rule(original_name, current_list, remove=True)
-
-                # If rule localized, enable on current list
-                if current_list == "ops":
-                    acl_object.op_player(acl_object.displayed_rule.rule)
-                elif current_list == "bans":
-                    acl_object.ban_player(acl_object.displayed_rule.rule)
-                elif current_list == "wl":
-                    acl_object.whitelist_player(acl_object.displayed_rule.rule)
-
-                hover_attr = (
-                    icon_path("earth-strike.png"), 'LOCALIZE',
-                    (0.439, 0.839, 1, 1) if current_list == "ops" else
-                    (1, 0.5, 0.65, 1) if current_list == "bans" else
-                    (0.3, 1, 0.6, 1)
-                )
-                banner_text = f"'${filtered_name}$' is now locally applied"
-                new_scope = "local"
-                reload_page = True
-
-            elif "globalize" in option:
-                acl_object.add_global_rule(original_name, current_list, remove=False)
-                hover_attr = (icon_path("earth-sharp.png"), 'GLOBALIZE', (0.953, 0.929, 0.38, 1))
-                banner_text = f"'${filtered_name}$' is now globally applied"
-                new_scope = "global"
-                reload_page = True
-
-
-        # Operator options
-        elif current_list == "ops":
-            if "demote" in option:
-                if self.displayed_scope == "global":
-                    acl_object.add_global_rule(original_name, current_list, remove=True)
-                else:
-                    acl_object.op_player(original_name, remove=True)
-
-                hover_attr = (icon_path("close-circle.png"), 'DEMOTE', (1, 0.5, 0.65, 1))
-                banner_text = f"'${filtered_name}$' was demoted"
-                new_scope = "local"
-                reload_page = True
-
-            elif "promote" in option:
-                acl_object.op_player(original_name, remove=False)
-                hover_attr = (icon_path("promote.png"), 'PROMOTE', (0.3, 1, 0.6, 1))
-                banner_text = f"'${filtered_name}$' was promoted"
-                reload_page = True
-
-
-        # Ban options Player/IP
-        elif current_list == "bans":
-            if acl_object.displayed_rule.rule_type == "player":
-                if "ban user" in option:
-                    acl_object.ban_player(original_name, remove=False)
-                    hover_attr = (icon_path("close-circle.png"), 'BAN', (1, 0.5, 0.65, 1))
-                    banner_text = f"'${filtered_name}$' is banned"
-                    reload_page = True
-
-                elif "ban IP" in option:
-                    acl_object.ban_player(ip_addr, remove=False)
-
-                    if self.displayed_scope == "global":
-                        acl_object.add_global_rule(original_name, current_list, remove=True)
-
-                    acl_object.ban_player(f"!w{ip_addr}", remove=True)
-                    hover_attr = (icon_path("close-circle.png"), 'BAN', (1, 0.5, 0.65, 1))
-                    banner_text = f"'${filtered_name}$' is banned"
-                    reload_page = True
-
-                if "pardon IP" in option and "user" in option:
-                    acl_object.ban_player([original_name, ip_addr], remove=True)
-
-                    if self.displayed_scope == "global":
-                        acl_object.add_global_rule(original_name, current_list, remove=True)
-
-                    # Whitelist IP if it's still in the rule list
-                    if ip_addr in acl.gen_iplist(acl_object.rules['subnets']):
-                        acl_object.ban_player(f"!w{ip_addr}", remove=False)
-
-                    hover_attr = (icon_path("lock-open.png"), 'PARDON', (0.3, 1, 0.6, 1))
-                    banner_text = f"'${filtered_name}$' is pardoned"
-                    new_scope = "local"
-                    reload_page = True
-
-                elif "pardon user" in option:
-                    if self.displayed_scope == "global":
-                        acl_object.add_global_rule(original_name, current_list, remove=True)
-                    else:
-                        acl_object.ban_player(original_name, remove=True)
-
-                    hover_attr = (icon_path("lock-open.png"), 'PARDON', (0.3, 1, 0.6, 1))
-                    banner_text = f"'${filtered_name}$' is pardoned"
-                    new_scope = "local"
-                    reload_page = True
-
-                elif "pardon IP" in option:
-                    acl_object.ban_player(ip_addr, remove=True)
-
-                    # Whitelist IP if it's still in the rule list
-                    if ip_addr in acl.gen_iplist(acl_object.rules['subnets']):
-                        acl_object.ban_player(f"!w{ip_addr}", remove=False)
-
-                    hover_attr = (icon_path("lock-open.png"), 'PARDON', (0.3, 1, 0.6, 1))
-                    banner_text = f"'${filtered_name}$' is pardoned"
-                    new_scope = "local"
-                    reload_page = True
-
-
-            # IP rules
-            else:
-                if "ban" in option:
-                    # If rule is global and ban is added, switch scope to local
-                    if "whitelist" in acl_object.displayed_rule.display_data['rule_info'] and self.displayed_scope == "global":
-                        new_scope = "local"
-
-                    if "!w" in original_name:
-                        acl_object.ban_player(original_name, remove=True)
-                        acl_object.ban_player(filtered_name, remove=False)
-                    else:
-                        acl_object.ban_player(original_name, remove=False)
-                    hover_attr = (icon_path("close-circle.png"), 'BAN', (1, 0.5, 0.65, 1))
-                    banner_text = f"'${filtered_name}$' is banned"
-                    new_name = filtered_name
-                    reload_page = True
-
-                elif "pardon" in option:
-                    if self.displayed_scope == "global":
-                        acl_object.add_global_rule(original_name, current_list, remove=True)
-                    else:
-                        acl_object.ban_player(original_name, remove=True)
-
-                    hover_attr = (icon_path("lock-open.png"), 'PARDON', (0.3, 1, 0.6, 1))
-                    banner_text = f"'${filtered_name}$' is pardoned"
-                    new_scope = "local"
-                    reload_page = True
-
-                elif "remove" in option:
-                    if self.displayed_scope == "global":
-                        acl_object.add_global_rule(original_name, current_list, remove=True)
-                    else:
-                        acl_object.ban_player(original_name, remove=True)
-
-                    hover_attr = (icon_path("shield-disabled-outline.png"), 'REMOVE', (0.7, 0.7, 1, 1))
-                    banner_text = f"'${filtered_name}$' was removed"
-                    new_scope = "local"
-                    reload_page = True
-
-                elif "whitelist" in option:
-                    # If rule is global and whitelist is added, switch scope to local
-                    if "ban" in acl_object.displayed_rule.display_data['rule_info'] and self.displayed_scope == "global":
-                        new_scope = "local"
-
-                    acl_object.ban_player(original_name, remove=True)
-                    acl_object.ban_player(f"!w{filtered_name}", remove=False)
-                    hover_attr = (icon_path("shield-checkmark-outline.png"), 'WHITELIST', (0.439, 0.839, 1, 1))
-                    banner_text = f"'${filtered_name}$' is whitelisted"
-                    new_name = f"!w{filtered_name}"
-                    reload_page = True
-
-
-        # Whitelist options
-        elif current_list == "wl":
-            if "restrict" in option:
-                if self.displayed_scope == "global":
-                    acl_object.add_global_rule(original_name, current_list, remove=True)
-                else:
-                    acl_object.whitelist_player(original_name, remove=True)
-
-                hover_attr = (icon_path("close-circle.png"), 'RESTRICT', (1, 0.5, 0.65, 1))
-                banner_text = f"'{filtered_name}' is restricted"
-                new_scope = "local"
-                reload_page = True
-
-            elif "permit" in option:
-                acl_object.whitelist_player(original_name, remove=False)
-                hover_attr = (icon_path("checkmark-circle-sharp.png"), 'PERMIT', (0.3, 1, 0.6, 1))
-                banner_text = f"'{filtered_name}' is permitted"
-                reload_page = True
-
-
-        if reload_page:
-
-            utility.screen_manager.current_screen.update_list(current_list, reload_children=False)
-
-            Clock.schedule_once(
-                functools.partial(
-                    utility.screen_manager.current_screen.show_banner,
-                    hover_attr[2],
-                    banner_text,
-                    hover_attr[0],
-                    2,
-                    {"center_x": 0.5, "center_y": 0.965}
-                ), 0
-            )
-
-            def trigger_highlight(*args):
-                for rule_button in utility.screen_manager.current_screen.scroll_layout.children:
-
-                    if rule_button.rule.rule == original_name:
-                        rule_button.highlight(rule_button.button.background_color, rule_button.text.color, hover_attr[2])
-
-                    else:
-                        rule_button.button.on_leave()
-                        Animation.cancel_all(rule_button.button)
-                        Animation.cancel_all(rule_button.text)
-                        Animation.cancel_all(rule_button.icon)
-                        rule_button.change_properties(rule_button.rule)
-
-                    rule_button.button.ignore_hover = False
-
-            Clock.schedule_once(trigger_highlight, 0)
-
-        # Update display rule regardless of button pressed
-        utility.screen_manager.current_screen.update_user_panel(new_name, new_scope)
-
-        # print(option)
-
-
 class AddonButton(HoverButton):
 
     def toggle_installed(self, installed, *args):
@@ -9726,132 +8333,3 @@ class AddonButton(HoverButton):
         else:
             self.subtitle.text = self.original_subtitle
             self.subtitle.font_name = self.original_font
-
-
-# Create demo of how the server will appear in the Server Manager:
-class ServerDemoInput(BaseInput):
-
-    def resize_self(self, *args):
-        offset = 9.45 if self.type_image.type_label.text in ["vanilla", "paper", "purpur"]\
-            else 9.6 if self.type_image.type_label.text == "forge"\
-            else 9.35 if self.type_image.type_label.text == "craftbukkit"\
-            else 9.55
-
-        self.type_image.image.x = self.width+self.x-self.type_image.image.width-self.padding_x[0]+10
-        self.type_image.image.y = self.y+(self.padding_y[0]/2.7)
-
-        self.title_t.pos = (self.x + 185, self.y - 7)
-
-        # Telepath icon
-        if self.type_image.tp_shadow:
-            self.type_image.tp_shadow.pos = (self.type_image.image.x - 2, self.type_image.image.y)
-            self.type_image.tp_icon.pos = (self.type_image.image.x - 2, self.type_image.image.y)
-
-        self.type_image.type_label.x = (self.width+self.x-(self.padding_x[0]*offset)) - 3
-        self.type_image.type_label.y = self.y+(self.padding_y[0]/6.9)
-
-        self.type_image.version_label.x = (self.width+self.x-(self.padding_x[0]*offset)) - 3
-        self.type_image.version_label.y = self.y-(self.padding_y[0]*0.85)
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-        self.halign = "left"
-        self.properties = {"type": "", "version": "", "name": ""}
-        self.padding_x = 30
-        self.padding_y = 24.5
-        self.font_size = sp(25)
-        self.size_hint_max = (580, 80)
-        self.hint_text_color = (0.65, 0.65, 1, 1)
-        self.background_normal = os.path.join(paths.ui_assets, 'server_preview.png')
-        self.title_text = ""
-        self.hint_text = ""
-        self.markup = True
-
-
-        # Type icon and info
-        with self.canvas.after:
-            self.type_image = RelativeLayout()
-            self.type_image.image = Image(source=None)
-            self.type_image.image.allow_stretch = True
-            self.type_image.image.size = (62, 62)
-            self.type_image.image.color = (0.65, 0.65, 1, 1)
-
-            def TemplateLabel():
-                template_label = AlignLabel()
-                template_label.halign = "right"
-                template_label.valign = "middle"
-                template_label.text_size = template_label.size
-                template_label.font_size = sp(18)
-                template_label.color = self.foreground_color
-                template_label.font_name = self.font_name
-                template_label.width = 200
-                return template_label
-
-            self.type_image.version_label = TemplateLabel()
-            self.type_image.version_label.color = (0.6, 0.6, 1, 0.6)
-            self.type_image.type_label = TemplateLabel()
-            self.type_image.type_label.font_size = sp(22)
-            self.type_image.tp_shadow = None
-
-            self.bind(pos=self.resize_self)
-
-            self.title_t = AlignLabel(halign='left', valign='center')
-            self.title_t.font_size = sp(25)
-            self.title_t.size_hint_max = (400, 80)
-            self.title_t.text_size = self.title_t.size_hint_max
-            self.title_t.color = (0.65, 0.65, 1, 1)
-            self.title_t.markup = True
-            self.title_t.font_name = self.font_name
-            self.add_widget(self.title_t)
-
-    # Make the text box non-interactive
-    def on_enter(self, value):
-        return
-
-    def on_touch_down(self, touch):
-        self.focus = False
-
-    def keyboard_on_key_down(self, window, keycode, text, modifiers):
-        return
-
-    def insert_text(self, substring, from_undo=False):
-        return
-def server_demo_input(pos_hint, properties):
-    demo_input = ServerDemoInput()
-    demo_input.properties = properties
-    demo_input.pos_hint = pos_hint
-    demo_input.__translate__ = False
-    demo_input.title_t.text = properties['name']
-    demo_input.type_image.version_label.__translate__ = False
-    demo_input.type_image.version_label.text = properties['version']
-    demo_input.type_image.type_label.__translate__ = False
-    demo_input.type_image.type_label.text = properties['type'].lower().replace("craft", "")
-    demo_input.type_image.image.source = os.path.join(paths.ui_assets, 'icons', 'big', f'{properties["type"].lower()}_small.png')
-
-    if utility.screen_manager.current.startswith('CreateServer'): send_log('CreateServer', f"menu progress:\n{properties}", 'info')
-    if properties['_telepath_data']:
-        if properties['_telepath_data']['nickname']:
-            head = properties['_telepath_data']['nickname']
-        else:
-            properties['_telepath_data']['host']
-
-        demo_input.title_t.text = f"[color=#7373A2]{head}/[/color]{properties['name']}"
-
-        with demo_input.canvas.after:
-            demo_input.type_image.tp_shadow = Image(source=icon_path('shadow.png'))
-            demo_input.type_image.tp_shadow.allow_stretch = True
-            demo_input.type_image.tp_shadow.size_hint_max = (33, 33)
-            demo_input.type_image.tp_shadow.color = constants.background_color
-            demo_input.type_image.add_widget(demo_input.type_image.tp_shadow)
-
-            demo_input.type_image.tp_icon = Image(source=icon_path('telepath.png'))
-            demo_input.type_image.tp_icon.allow_stretch = True
-            demo_input.type_image.tp_icon.size_hint_max = (33, 33)
-            demo_input.type_image.tp_icon.color = demo_input.type_image.image.color
-            demo_input.type_image.add_widget(demo_input.type_image.tp_icon)
-
-    return demo_input
-
-
-
